cisco: move context execution from cisco.manager to cisco.client
enumerate_contexts(), MultiContextExecutor, CiscoFwContext are really
more part of the client anyway.

This also adds the ability to listen for connection_made and
connection_lost events from CiscoSSHClient which fixes a boog whereby
MultiContextExecutor would not isssue a changeto command after a
disconnect/reconnect, thus potentially causing commands to be executed
in the wrong context.